refactor: blockidier regression test rename test files
