+++ crates/blockifier_regression_test/resources/raw_rpc_json_objects/deprecated_contract_class.json
