+++ crates/blockifier_regression_test/src/lib.rs
@@ -1,3 +1,1 @@
-#[cfg(test)]
-pub mod json_test;
 pub mod state_reader;
