+++ crates/blockifier_regression_test/src/state_reader.rs
@@ -1,6 +1,8 @@
 pub mod compile;
 #[cfg(test)]
+pub mod raw_rpc_json_test;
+#[cfg(test)]
 #[cfg(feature = "blockifier_regression_https_testing")]
-pub mod rpc_test;
+pub mod rpc_https_test;
 pub mod test_state_reader;
 pub mod utils;
